+++ ErsatzTV.Infrastructure.MySql/Migrations/20240209191908_Reset_SongMetadataBadMigration.cs
@@ -11,10 +11,12 @@ namespace ErsatzTV.Infrastructure.MySql.Migrations
         protected override void Up(MigrationBuilder migrationBuilder)
         {
             migrationBuilder.Sql(@"UPDATE SongMetadata SET Artists = '[]', AlbumArtists = '[]'");
-            
+
             migrationBuilder.Sql(
                 @"UPDATE LibraryPath SET LastScan = '0001-01-01 00:00:00' WHERE Id IN
-                (SELECT LP.Id FROM LibraryPath LP INNER JOIN Library L on L.Id = LP.LibraryId WHERE MediaKind = 5)");
+                (SELECT A.lpid FROM
+                (SELECT LP.Id AS lpid FROM LibraryPath LP INNER JOIN Library L on L.Id = LP.LibraryId WHERE MediaKind = 5)
+                AS A)");
 
             migrationBuilder.Sql(
                 @"UPDATE Library SET LastScan = '0001-01-01 00:00:00' WHERE MediaKind = 5");
@@ -24,7 +26,9 @@ namespace ErsatzTV.Infrastructure.MySql.Migrations
             
             migrationBuilder.Sql(
                 @"UPDATE LibraryFolder SET Etag = NULL WHERE Id IN
-                (SELECT LF.Id FROM LibraryFolder LF INNER JOIN LibraryPath LP on LF.LibraryPathId = LP.Id INNER JOIN Library L on LP.LibraryId = L.Id WHERE MediaKind = 5)");
+                (SELECT A.lfid FROM
+                (SELECT LF.Id AS lfid FROM LibraryFolder LF INNER JOIN LibraryPath LP on LF.LibraryPathId = LP.Id INNER JOIN Library L on LP.LibraryId = L.Id WHERE MediaKind = 5)
+                AS A)");
         }
 
         /// <inheritdoc />